+++ pkg/utils/utils.go
@@ -65,7 +65,7 @@ func IsValidProviderVersionFormat(version string) bool {
 }
 
 func IsValidProviderDocumentType(providerDocumentType string) bool {
-	validTypes := []string{"resources", "data-sources", "functions", "guides", "overview"}
+	validTypes := []string{"resources", "data-sources", "functions", "guides", "overview", "actions", "list-resources"}
 	return slices.Contains(validTypes, providerDocumentType)
 }
 
@@ -79,7 +79,7 @@ func LogAndReturnError(logger *log.Logger, context string, err error) error {
 }
 
 func IsV2ProviderDocumentType(dataType string) bool {
-	v2Categories := []string{"guides", "functions", "overview", "actions"}
+	v2Categories := []string{"guides", "functions", "overview", "actions", "list-resources"}
 	return slices.Contains(v2Categories, dataType)
 }